Add timer messages
These are messages that have a duration and will be scheduled to be
sent back to the component itself after that duration has occured.
The idea is that most uses of ticks are really that you want to wait
for a certain amount of time and then check again for some condition.
Modelling it this way we will be independent of tick-rate changes, and
the scheduler will move at the speed of messages (rather than also having
to work at a fixed clock-speed for ticks).

At the moment these messages occurs during a logical time, but does
not get appended to the network trace. This will require a change to the
table so that ldfi can exclude these messages.