Update the debugger,
Rather than displaying all components, we now have a list of all components
that you can select which one you want to see the state(-diff) for. When
changing to a new event, it will automatically select the component that
received the message (if it was a component).

Internal: I added a `DebugApplication` type to keep track of all the state
and other information that the application is using. There is still some
spaghetti dependencies but I think it is a bit better.