feat(executor+scheduler): Add MetaInfo for each execution step
When Scheduler sends information to executor (Event, Timer) it now
also has a field in the JSON with meta-information. This meta-information
currently contains:

```
{"test-id": <current-test-id>,
 "run-id": <current-run-id>,
 "logical-time": <current-logical-time>}
```

This information should not be sent to the SUT, but only used by the executor
when creating events for the event-log, to tag the appropriate meta information.

This will allow further refactorings of the executor to not have to know the
TestId, RunId. And should enable the executor in the future to be able to
send `Info` events in a straight-forward way.

This commit also changes the default way we (un)marshal TestId and RunId to/from
JSON. It now only uses the underlying number directly, and in the few cases we want
to have a wrapper object, we have to make such a wrapper ourselves.