fix(executor): ScheduledEvent was not parsing meta field properly
The unmarshaling of `ScheduledEvent` happens in `UnmarshalScheduledEvent`
in lib/marshaler.go and this function was not updated to have the
`MetaInfo` field. This is now fixed.

I also removed the json-annotations from `ScheduledEvent` since they
are not actually used, and can be confusing and tricking you to
beleive they are all there is to unmarshal the type.