feat(executor): Executor now sends the ExecutionStep event
This event unifies the old `log_trace` table and `heap_trace` table
in one event. This pr now makes it so the executor only calls one function
and it will emit this event, and also append to the old tables for now.
In the future we will be able to remove those appends.

This also allowed some more refactoring of the handlers, we no longer use the
`EventLogEmitter` so this could be removed. This also simplied the `Executor`
type.