refactor(checker): Use the new NetworkTrace and introduce jepsen_history view
This commit introduces the `jepsen_history` view that is used instead of the old
`history` table by the checker. This also required a small change to the
scheduler when it emits the `NetworkTrace` event for messages that are sent back
to the client. There was some special code to remove one layer of `response`
field from the `args`. Now this stripping is happening in the entire event,
rather than just do it for the `history` only. Less special casing.

Now nothing depends on the `history` table and is therefore removed.