refactor: Remobe old network_trace and use a view
Removing the old tables `network_trace` and `time_mapping` and introduces a new
view `network_trace` that contains the combined information. I renamed some
columns in the view since naming a column `from` in sql is fraught with danger.

Also instead of using `at` we now use the `recv_logical_time` which in my mind
is more expressive of what this time represents.