refactor(executor): Prepare executor for CreateRun event
`lib.CreateRun` now also takes a struct with all options for the run, which will
soon become the `CreateRun` event. At the moment it will do what to old
`CreateRun` function did, and then call each separate end-point to set the
settings. But this will be updated later to be one call to scheduler.