bug(scheduler): Expire clients before we decide to delay message
We now expire client requests (i.e time them out) before we decide that we
should delay a message. This is because if the delayed message is the last thing
in the agenda, it would never be scheduled since we would never get to expire
the previous call, so it would get into an infinte loop and keep delaying.