feat(scheduler): add new create-run-event! endpoint in Scheduler
This endpoint will create a run, and set all the settings in one go, this is so
we in the future can have a CreateRun event in the event_log with all the
settings.

This endpoint should also be able to create the new event which should be able
to deprecate some older tables (such as `run` and `faults`, though `faults` will
need yet another event).