refactor: Use CreateTest event
The final push to make everything that goes to the db be an event to the
event_log.

This commit creates the `CreateTest` event, which contains the agenda and the
deployment. There is now also an `test_info` view that will contain the agenda
and deployment as columns (the data is still in json). This means that we no
longer use the `agenda`, `deployment` or `test` tables.

Noticeable renamings:
* in deployment we now call the reactors `reactor` instead of `component`.