perf(debugger): Implement our own sequence diagrams
Calling `plantuml` is really slow (probably because it is java application and
java boot is slow), so this commit replaces that with our own sequence diagram
drawing code. It should be reasonable feature complete with what we had before,
but it should take significantly less time to compute the diagrams, and
therefore also boot the debugger.

The diagrams are drawn in a lazy fashion, where we only draw the diagram if it
is looked at. This might not really be needed since we draw them so fast, but I
already had the code for when I was trying to get `plantuml` to work that I
decided to keep it.

closes #66