bug(executor): Logs for reactors were attributed to wrong handler
We collect the logs before we run the reactor handling an event, which of course
got us to mis-attribute the log messages. The reason this happened was that
before we didn't collect logs, but rather set the logger to output to the
correct meta information, but when this change happened we didn't move the
collection code (which used to be setup code) to after the reactor ran.