feat(ltl): Add new command for ltl to check many formulas
Added new command called `checkmany` which takes two arguments:
  --testId which is the test id to check
  --runId which is the run id to check
And it will also read the stdin, each line of stdin will be a different formula
to check. The stdout will be an json object with one field `results` which is an
json array of results (similar to the `check` command).