perf(debugger): Make the sequence diagram faster
tview is quite slow when giving large text (with lots of unicode) to print for
their TextView. This is exactly the scenario we have with the sequence diagram,
so now we will slice the text to only what would be visible in the window.

This change makes it much faster to scroll in large histories, but notice that
we no longer can scroll in the diagram buffer (but we can in the events buffer
which will update the diagram buffer).