feat(runtime): Executors can now restart the reactors
A new admin command has been added to restart the reactors. The admin command is
called `AdminResetReactors` and can for example be invoked like this:
```bash
$ echo "AdminResetReactors" | nc -U /tmp/executor-admin.sock
Reactors restarted
```

When creating an executor program, one now have to give a function creating a
topology, rather than just a topology. So when we reset, we just create a new
topology for the executor to run on.