refactor(runtime): New executor split up how we build topology
Previously we gave a function `func() lib.Topology` to get the topology for the
executor. Now we have split it up into one array `[]string` of the names of the
reactors, and a single constructor `func(name string) lib.Reactor`.

The reason for this change is so that the executor will be able to restart a
single reactor (for example to simulate an error).