feat(runtime): Executor now give a `LogWriter` to all reactors in the constructor
This `LogWriter` implements the `io.Writer` interface and also the `Sync()`
interface that are used by the `zap` logging framework (and potentially others).
The executor will make sure that each time a reactor logs, that the log messages
will be correctly associated together with the message the reactor is reacting
to.