experiment(journal): Add some TH to compute offsets
This is an experiment to automate the computation for offset of fields in a
record. This computation doesn't take alignment into account (and I'm not sure
it should). A benefit of having the macro is that when the type changes we don't
have to manually change the constants.

It works by adding a new macro `offsetTo` that given a field name computes the
size of all fields preceeding it.

```haskell
*Journal> :set -XTemplateHaskell
*Journal> import OffsetMacro
*Journal OffsetMacro> $(offsetTo 'mdTermLength)
32
*Journal OffsetMacro> lOG_TERM_LENGTH_OFFSET
32
```

At the moment it computes to an expression that will do the sum at runtime, this
probably will be optimized away by the compiler.

This do require Template Haskell to work, with all what that entails, so not
sure yet if it is worth it.