refactor(sut): Change the `Blocker` interface for dumblog
The new interface removes the need for the worker thread needing to do retries,
instead the frontend will first add the key to the map, and then add the event
on the journal. This commit adds a `Key` type that bundles the `Int` key and the
`MVar`.

Another change is that the `AtomicCounter` has been moved to the `Blocker` type
rather than being part of the `FrontEndInfo` state. It makes more sense to have
the counter here, since now we want to add to the map when creating keys
anyways.