feat(course): Add field if history event was dropped or not
The information if an event was dropped or not is not stored in the `HistEvent`
type but rather in a new wrapper type named `HistEvent'`. The reason for doing
it this way is that the `Deployment` type still only have:
```
dAppendHistory :: Deployment-> HistEvent -> IO ()
```
So `HistEvent` is part of the interface that the `eventLoop` uses. And the
eventloop should not now if events arrived or not.

Currently all events will eventually arrive in simulation, but when introducing
faults that may not be true. So in the future a faulty `Network` mock could add
to the history dropped messages.